pyproject.toml
[tool.poetry]
name = "mediaforge"
version = "1.0.0"
description = "A Discord bot for editing and creating videos, images, GIFs, and more!"
authors = ["reticivis"]
license = "GPL-3.0"

[tool.poetry.dependencies]
python = "^3.8"
nextcord = { git = "https://github.com/nextcord/nextcord.git", allow-prereleases = true, extras = ["speed"] }
"discordlists.py" = { git = "https://github.com/HexCodeFFF/discordlists.py.git" }
aiofiles = "^0.8.0"
humanize = "^3.13.1"
coloredlogs = "^15.0.1"
selenium = "^4.1.0"
Pillow = "^9.0.0"
numpy = "^1.22.1"
pyautotune = { git = "https://github.com/HexCodeFFF/PyAutoTune.git" }
regex = ">=2019.12.20"
pronouncing = "^0.2.0"
aiosqlite = "^0.17.0"
psutil = "^5.9.0"
aubio = "^0.4.9"
apng = "^0.3.4"
docstring-parser = "^0.13"
yt-dlp = ">=2022.1.21"
emojis = "^0.6.0"
python-magic = { version = "^0.4.13" }
python-magic-win64 = { version = "^0.4.13", platform = "win32" }
requests = "^2.27.1"
thefuzz = "^0.19.0"
python-Levenshtein = "^0.12.2"
cyhunspell = "^2.0.2"


[tool.poetry.dev-dependencies]
#aiohttp = { version = "^3.8.1", extras = ["speedups"] }
#setuptools = "^60.5.0"
isort = "^5.10.1"
poetry = "^1.1.12"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
